Chain of fixes on product svc/item limitations
The original issue (impacting OMT-170) exposed a bug in the validation of product limitations which in turn showed that in the tests, we should not create ProductItem or ProductService with a limitation type (mandatory) but no value. This is now supported in the code even though it should not happen in reality.

Requires a fix in the openimis-be-product_py module