OMT-208 validation could override review rejection
When a review rejected all the items/services, they were still passed through the validation and, as the validation was still ok, would be reset back to PASSED.
We fixed the validation to check only not yet rejected items and services.

All this is illustrated with the test checking that the dedrem is effectively deleted in this situation.