Refactor to allow FHIR to submit directly
FHIR was still using the legacy stored procedure to submit claims. To submit them through the Python code, we moved a lot of code from the migrations to the services.
However, since the original services where imported with * and that was needed, the services were split.
Importing from .services import * meant that all methods in services were exposed as methods of the package and with additional methods using the Model objects, that triggered a loading chain before the Model was available.